rename current_user_dependency to dependencies
